revise demo
make toggle block ‘buttons’ set to display inline-block instead of block.

revise class names and css selectors:
originally the idea here was that this could be the foundation for an accordion script. but as accordions have many other things to take into consideration, this script was reduced down to a single show/hide toggle component.  the class names and example markup was still setup for accordions though, so this has been revised to allow for more flexible setups.